refactor: App클래스에서 다리를 생성한 후, 게임을 시작하도록 설계 변경
